[add]日付変更とランキングの連携
