+++ src/app/weekly/page.jsx
@@ -7,6 +7,7 @@ import { Heatmap } from "@/components/charts/heatmap/weekly/Heatmap";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
+import { CalculateTotalRanking } from "@/features/calcWeekly";
 
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
@@ -15,107 +16,26 @@ dayjs.extend(utc);
 dayjs.extend(timezone);
 dayjs.tz.setDefault("Asia/Tokyo");
 
-// 週番号を取得する関数
-const getWeekNumber = (date) => {
-  const onejan = new Date(date.getFullYear(), 0, 1);
-  const weekNumber = Math.ceil(
-    ((date - onejan) / 86400000 + onejan.getDay() + 1) / 7,
-  );
-  return weekNumber;
-};
-
-// 滞在時間を計算する関数
-const calculateTotalTime = (checkIn, checkOut) => {
-  const checkInTime = new Date(checkIn).getTime();
-  const checkOutTime = new Date(checkOut).getTime();
-  const totalTime = checkOutTime - checkInTime;
-
-  return totalTime;
-};
-
-// ランキングデータを計算する関数
-const calculateWeeklyRanking = (data) => {
-  const weeklyRanking = {};
-
-  const weeklyData = data.reduce((acc, entry) => {
-    const weekNumber = getWeekNumber(new Date(entry.check_in));
-    acc[weekNumber] = acc[weekNumber] || [];
-    acc[weekNumber].push(entry);
-    return acc;
-  }, {});
-
-  for (const [weekNumber, entries] of Object.entries(weeklyData)) {
-    const weeklyRank = entries.reduce((acc, entry) => {
-      const studentId = entry.user_id;
-      const totalTime = calculateTotalTime(entry.check_in, entry.check_out);
-
-      acc[studentId] = (acc[studentId] || 0) + totalTime;
-      return acc;
-    }, {});
-
-    weeklyRanking[weekNumber] = Object.entries(weeklyRank)
-      .map(([studentId, totalTime]) => ({
-        studentId,
-        studentName: studentId,
-        totalTime,
-      }))
-      .sort((a, b) => b.totalTime - a.totalTime);
-  }
-
-  return weeklyRanking;
-};
-
-const getHourlyOccupancy = (data) => {
-  const hourlyOccupancy = {};
-
-  data.forEach((entry) => {
-    const checkInTime = new Date(entry.check_in);
-    const checkOutTime = new Date(entry.check_out);
-    const dateKey = checkInTime.toISOString().split("T")[0];
-
-    // もし指定の日付がまだオブジェクトに存在しない場合、初期化
-    if (!hourlyOccupancy[dateKey]) {
-      hourlyOccupancy[dateKey] = {};
-      for (let hour = 0; hour < 24; hour++) {
-        hourlyOccupancy[dateKey][hour] = 0;
-      }
-    }
-
-    // 部屋にいる時間を1時間ごとに分割
-    let currentTime = new Date(checkInTime);
-    while (currentTime < checkOutTime) {
-      const hour = currentTime.getHours();
-      hourlyOccupancy[dateKey][hour] += 1;
-
-      currentTime.setHours(currentTime.getHours() + 1);
-    }
-  });
-
-  return hourlyOccupancy;
-};
-
 export default function WeeklyPage() {
   const [weeklyRank, setWeeklyRank] = useState(null);
-  const [hourlyOccupancy, setHourlyOccupancy] = useState(null);
   const [heatmapData, setHeatmapData] = useState(null);
-  const [pickerTime, setPickerTime] = useState(null);
-
-  const today = dayjs().tz();
+  const [pickerTime, setPickerTime] = useState(dayjs());
 
   useEffect(() => {
     const fetchData = async () => {
       try {
         const response = await fetch(
-          // `/api/access/timestamp/${today.format("YYYY/MM")}`,
-          `/api/access/timestamp/2023/12`,
+          `/api/access/timestamp/${pickerTime.format("YYYY/MM")}`,
         );
         const result = await response.json();
 
-        const weeklyRankData = calculateWeeklyRanking(result);
-        const hourlyOccupancyData = getHourlyOccupancy(result);
+        const oneWeekAgo = pickerTime.subtract(1, "week").startOf("day");
+        const filteredData = result.filter((entry) =>
+          dayjs(entry.check_in).isBetween(oneWeekAgo, pickerTime),
+        );
 
-        setWeeklyRank(weeklyRankData[50]);
-        setHourlyOccupancy(Object.values(hourlyOccupancyData));
+        const weeklyRankData = CalculateTotalRanking(filteredData);
+        setWeeklyRank(weeklyRankData);
       } catch (error) {
         console.log(error);
       }
@@ -128,11 +48,13 @@ export default function WeeklyPage() {
       }
     };
     fetchData();
-  }, []);
+  }, [pickerTime]);
 
   const onPickerChange = (newDate) => {
     setPickerTime(newDate);
   };
+
+  console.log(pickerTime);
 
   return (
     <Grid container spacing={2}>
@@ -152,7 +74,7 @@ export default function WeeklyPage() {
         </Grid>
         <Grid item xs={12}>
           <Text message="時間帯別滞在人数" />
-          <Heatmap hourlyOccupancy={hourlyOccupancy} data={heatmapData} />
+          <Heatmap data={heatmapData} />
         </Grid>
       </Grid>
